Add hook 'useConnect'
Eliminates the need for 'WalletConnectionButton' in the sample dapp.
