+++ packages/react-components/src/index.ts
@@ -1,3 +1,4 @@
+export * from './useConnect';
 export * from './useConnection';
 export * from './useContractSelector';
 export * from './useWalletConnectorSelector';
